Rename metadata.py to __metadata__.py
