+++ pydantic_argparse/__init__.py
@@ -10,7 +10,7 @@ The public interface exposed by this package is the declarative and typed
 
 # Local
 from .argparse import ArgumentParser
-from .metadata import __title__, __description__, __version__, __author__, __license__
+from .__metadata__ import __title__, __description__, __version__, __author__, __license__
 
 # Public Re-Exports
 __all__ = (
