+++ pydantic_argparse/__metadata__.py
